LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/not-in-shadow-tree-expected.txt

CONSOLE MESSAGE: line 3463: Error: assert_unreached: Got more than 1 load event Reached unreachable code

Harness Error (FAIL), message = Error: assert_unreached: Got more than 1 load event Reached unreachable code
